server/requirements.txt
# FastAPI and Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
starlette>=0.27.0

# Database and ORM
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Data Validation and Settings
pydantic>=2.5.0
pydantic-settings>=2.0.3
email-validator>=2.0.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Environment and Configuration
python-dotenv>=1.0.0

# Machine Learning and AI
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.35.2
sentence-transformers>=2.2.2
Pillow>=10.1.0

# Text Processing and OCR
pytesseract==0.3.10
opencv-python==4.8.1.78

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# File Handling
python-multipart==0.0.6

# Logging and Monitoring
loguru==0.7.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Utilities
tqdm==4.66.1
numpy==1.24.3
pandas==2.1.3

# Optional: For advanced NLP features
# spacy==3.7.2
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
